Conflicts:
	Makefile

+++ build/post-to-upgrade.sh
@@ -8,10 +8,11 @@
 
 # It's possible these should be set via getopt.
 
+: ${NANO_LABEL:=FreeNAS}
 : ${UPDATE_HOST:=beta-update.freenas.org}
 : ${UPDATE_USER:=jkh}
-: ${UPDATE_DB:="sqlite:updates.db"}
-: ${UPDATE_DEST:=/tank/www/FreeNAS}
+: ${UPDATE_DB:="sqlite:${NANO_LABEL}-updates.db"}
+: ${UPDATE_DEST:=/tank/www/${NANO_LABEL}}
 
 prog=$(basename $0)
 usage() {
@@ -24,7 +25,7 @@ if [ $# -ne 1 ]; then
 fi
 
 SOURCE="$1"
-MKREL="/usr/local/bin/freenas-release -D ${UPDATE_DB} --archive ${UPDATE_DEST} add"
+MKREL="/usr/local/bin/freenas-release -P ${NANO_LABEL} -D ${UPDATE_DB} --archive ${UPDATE_DEST} add"
 
 set -e
 set -x
@@ -35,9 +36,16 @@ do
     test -f ${note} && cp ${note} "${SOURCE}"
 done
 
-TEMP_DEST=$(ssh ${UPDATE_USER}@${UPDATE_HOST} mktemp -d /tmp/update-XXXXXXXXX)
-scp -r "${SOURCE}/." ${UPDATE_USER}@${UPDATE_HOST}:${TEMP_DEST}
-ssh ${UPDATE_USER}@${UPDATE_HOST} "${MKREL} ${TEMP_DEST}"
+set -e
+TEMP_DEST=$(ssh ${UPDATE_USER}@${UPDATE_HOST} mktemp -d /tmp/update-${NANO_LABEL}-XXXXXXXXX)
+if [ $? -ne 0 -o -z "${TEMP_DEST}" ]; then
+    echo Cannot create temporary directory 1>&2
+    exit 1
+fi
+scp -r "${SOURCE}/." ${UPDATE_USER}@${UPDATE_HOST}:${TEMP_DEST} && \
+    ssh ${UPDATE_USER}@${UPDATE_HOST} "${MKREL} ${TEMP_DEST}"
 ssh ${UPDATE_USER}@${UPDATE_HOST} "rm -rf ${TEMP_DEST}"
+
+set +e
 
 exit 0
